# swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"IRegisterUserParams": {
			"properties": {
				"email": {
					"type": "string"
				},
				"password": {
					"type": "string"
				},
				"name": {
					"type": "string"
				}
			},
			"required": [
				"email",
				"password",
				"name"
			],
			"type": "object"
		},
		"ILoginParams": {
			"properties": {
				"email": {
					"type": "string"
				},
				"password": {
					"type": "string"
				}
			},
			"required": [
				"email",
				"password"
			],
			"type": "object"
		}
	},
	"info": {
		"title": "codinova",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "object"
						}
					}
				},
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "object"
						}
					}
				},
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "createUserParams",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IRegisterUserParams"
						}
					}
				]
			}
		},
		"/users/login": {
			"post": {
				"operationId": "LoginUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "object"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "loginParams",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ILoginParams"
						}
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"Bearer": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	}
}